Actualizacion de endpoints cristian

+++ api_SIF/Models/EmpleadosN/vacacione.cs
@@ -26,7 +26,7 @@ namespace api_SIF.Models.EmpleadosN
         [Column(TypeName = "int(11)")]
         public int? cantidadPago { get; set; }
         [Column(TypeName = "int(1)")]
-        public ulong? status { get; set; }
+        public int? status { get; set; }
 
 
         //[ForeignKey("id_empleado")]
